app/clean/requirements.txt
python-dotenv
mysql-connector-python
requests
datetime
timedelta